GH-291: Rename to MainThread for threading calls (#325)
* Rename to MainThread for threading calls

* Fix build issues
