+++ Samples/Samples/ViewModel/OrientationSensorViewModel.cs
@@ -110,7 +110,7 @@ namespace Samples.ViewModel
             {
                 case SensorSpeed.Fastest:
                 case SensorSpeed.Game:
-                    Platform.BeginInvokeOnMainThread(() =>
+                    MainThread.BeginInvokeOnMainThread(() =>
                     {
                         X = data.Orientation.X;
                         Y = data.Orientation.Y;
