+++ Xamarin.Essentials/MainThread/MainThread.android.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using Android.OS;
+
+namespace Xamarin.Essentials
+{
+    public static partial class MainThread
+    {
+        static Handler handler;
+
+        static bool PlatformIsMainThread
+        {
+            get
+            {
+                if (Platform.HasApiLevel(BuildVersionCodes.M))
+                    return Looper.MainLooper.IsCurrentThread;
+
+                return Looper.MyLooper() == Looper.MainLooper;
+            }
+        }
+
+        static void PlatformBeginInvokeOnMainThread(Action action)
+        {
+            if (IsMainThread)
+            {
+                action();
+                return;
+            }
+
+            if (handler?.Looper != Looper.MainLooper)
+                handler = new Handler(Looper.MainLooper);
+
+            handler.Post(action);
+        }
+    }
+}
