+++ Xamarin.Essentials/Platform/Platform.android.cs
@@ -14,7 +14,6 @@ namespace Xamarin.Essentials
 {
     public static partial class Platform
     {
-        static Handler handler;
         static ActivityLifecycleContextListener lifecycleListener;
 
         internal static Context AppContext =>
@@ -58,31 +57,6 @@ namespace Xamarin.Essentials
 
         internal static bool HasApiLevel(BuildVersionCodes versionCode) =>
             (int)Build.VERSION.SdkInt >= (int)versionCode;
-
-        static bool PlatformIsMainThread
-        {
-            get
-            {
-                if (HasApiLevel(BuildVersionCodes.M))
-                    return Looper.MainLooper.IsCurrentThread;
-
-                return Looper.MyLooper() == Looper.MainLooper;
-            }
-        }
-
-        static void PlatformBeginInvokeOnMainThread(Action action)
-        {
-            if (IsMainThread)
-            {
-                action();
-                return;
-            }
-
-            if (handler?.Looper != Looper.MainLooper)
-                handler = new Handler(Looper.MainLooper);
-
-            handler.Post(action);
-        }
 
         internal static CameraManager CameraManager =>
             AppContext.GetSystemService(Context.CameraService) as CameraManager;
