+++ Xamarin.Essentials/Platform/Platform.uwp.cs
@@ -1,25 +1,6 @@
-﻿using System;
-using Windows.ApplicationModel.Core;
-using Windows.UI.Core;
-
-namespace Xamarin.Essentials
+﻿namespace Xamarin.Essentials
 {
     public static partial class Platform
     {
-        static bool PlatformIsMainThread =>
-            CoreApplication.MainView.CoreWindow?.Dispatcher?.HasThreadAccess ?? false;
-
-        static void PlatformBeginInvokeOnMainThread(Action action)
-        {
-            var dispatcher = CoreApplication.MainView.CoreWindow?.Dispatcher;
-
-            if (dispatcher == null)
-                throw new InvalidOperationException("Unable to find main thread.");
-
-            if (!dispatcher.HasThreadAccess)
-                dispatcher.RunAsync(CoreDispatcherPriority.Normal, () => action()).WatchForError();
-            else
-                action();
-        }
     }
 }
