+++ docs/en/FrameworksIndex/xamarin-essentials.xml
@@ -241,6 +241,10 @@
     <Type Name="Xamarin.Essentials.MagnetometerData" Id="T:Xamarin.Essentials.MagnetometerData">
       <Member Id="P:Xamarin.Essentials.MagnetometerData.MagneticField" />
     </Type>
+    <Type Name="Xamarin.Essentials.MainThread" Id="T:Xamarin.Essentials.MainThread">
+      <Member Id="M:Xamarin.Essentials.MainThread.BeginInvokeOnMainThread(System.Action)" />
+      <Member Id="P:Xamarin.Essentials.MainThread.IsMainThread" />
+    </Type>
     <Type Name="Xamarin.Essentials.NetworkAccess" Id="T:Xamarin.Essentials.NetworkAccess">
       <Member Id="F:Xamarin.Essentials.NetworkAccess.ConstrainedInternet" />
       <Member Id="F:Xamarin.Essentials.NetworkAccess.Internet" />
@@ -285,10 +289,7 @@
       <Member Id="P:Xamarin.Essentials.Placemark.SubThoroughfare" />
       <Member Id="P:Xamarin.Essentials.Placemark.Thoroughfare" />
     </Type>
-    <Type Name="Xamarin.Essentials.Platform" Id="T:Xamarin.Essentials.Platform">
-      <Member Id="M:Xamarin.Essentials.Platform.BeginInvokeOnMainThread(System.Action)" />
-      <Member Id="P:Xamarin.Essentials.Platform.IsMainThread" />
-    </Type>
+    <Type Name="Xamarin.Essentials.Platform" Id="T:Xamarin.Essentials.Platform" />
     <Type Name="Xamarin.Essentials.Preferences" Id="T:Xamarin.Essentials.Preferences">
       <Member Id="M:Xamarin.Essentials.Preferences.Clear" />
       <Member Id="M:Xamarin.Essentials.Preferences.Clear(System.String)" />
