+++ docs/en/Xamarin.Essentials/Platform.xml
@@ -17,29 +17,6 @@
     </remarks>
   </Docs>
   <Members>
-    <Member MemberName="BeginInvokeOnMainThread">
-      <MemberSignature Language="C#" Value="public static void BeginInvokeOnMainThread (Action action);" />
-      <MemberSignature Language="ILAsm" Value=".method public static hidebysig void BeginInvokeOnMainThread(class System.Action action) cil managed" />
-      <MemberSignature Language="DocId" Value="M:Xamarin.Essentials.Platform.BeginInvokeOnMainThread(System.Action)" />
-      <MemberType>Method</MemberType>
-      <AssemblyInfo>
-        <AssemblyVersion>1.0.0.0</AssemblyVersion>
-        <AssemblyName>Xamarin.Essentials</AssemblyName>
-      </AssemblyInfo>
-      <ReturnValue>
-        <ReturnType>System.Void</ReturnType>
-      </ReturnValue>
-      <Parameters>
-        <Parameter Name="action" Type="System.Action" />
-      </Parameters>
-      <Docs>
-        <param name="action">Action to execute.</param>
-        <summary>Invokes an action on the main thread of the application.</summary>
-        <remarks>
-          <para></para>
-        </remarks>
-      </Docs>
-    </Member>
     <Member MemberName="Init">
       <MemberSignature Language="C#" Value="public static void Init (Android.App.Application application);" />
       <MemberSignature Language="ILAsm" Value=".method public static hidebysig void Init(class Android.App.Application application) cil managed" />
@@ -85,26 +62,6 @@
         <summary>Initialize Xamarin.Essentials with Android's activity and bundle.</summary>
         <remarks>
           <para></para>
-        </remarks>
-      </Docs>
-    </Member>
-    <Member MemberName="IsMainThread">
-      <MemberSignature Language="C#" Value="public static bool IsMainThread { get; }" />
-      <MemberSignature Language="ILAsm" Value=".property bool IsMainThread" />
-      <MemberSignature Language="DocId" Value="P:Xamarin.Essentials.Platform.IsMainThread" />
-      <MemberType>Property</MemberType>
-      <AssemblyInfo>
-        <AssemblyName>Xamarin.Essentials</AssemblyName>
-        <AssemblyVersion>1.0.0.0</AssemblyVersion>
-      </AssemblyInfo>
-      <ReturnValue>
-        <ReturnType>System.Boolean</ReturnType>
-      </ReturnValue>
-      <Docs>
-        <summary>Gets if it is the current main UI thread.</summary>
-        <value>If main thread.</value>
-        <remarks>
-          <para />
         </remarks>
       </Docs>
     </Member>
